fix(library): make every edge popover fully reactive and accessible
The compiler-reactivity fix had only reached the class-diagram edge
popover. Five other edge popovers (BPMN, Communication, Component,
Deployment, Use Case) still read their source/target node names through
imperative getNode() calls during render, so a swap — or a collaborator
renaming an endpoint — left the popover showing stale names once the React
Compiler memoized it. The swap control was also added unevenly: a labelled
button on the class popover, but a bare mouse-only <svg onClick> on the
other eight.

- Read endpoint names via the reactive useReactiveNodeName hook everywhere,
  closing the remaining staleness gap
- Extract a shared, keyboard-operable SwapEndsButton (native IconButton)
  that replaces nine duplicated swap controls — one accessible pattern,
  72 fewer lines
- Make the edge toolbar Edit/Delete controls native IconButtons so they are
  focusable and operable by keyboard, not merely announced as buttons
- Move the reactive single-element hooks out of the edge-named useEdges.ts
  into a cohesive useReactiveElement.ts
- Use slotProps.htmlInput instead of the deprecated MUI inputProps
