+++ library/lib/components/popovers/edgePopovers/ComponentDiagramEdgeEditPopover.tsx
@@ -1,11 +1,11 @@
 import { Box } from "@mui/material"
 import { EdgeStyleEditor, Typography } from "@/components/ui"
 import { useReactFlow } from "@xyflow/react"
-import { SwapHorizIcon } from "@/components/Icon"
-import { useEdgePopOver, useReactiveEdge } from "@/hooks"
+import { useEdgePopOver, useReactiveEdge, useReactiveNodeName } from "@/hooks"
 import { PopoverProps } from "../types"
 import { CustomEdgeProps } from "@/edges"
 import { EdgeTypeSelect, EdgeTypeOption } from "./EdgeTypeSelect"
+import { SwapEndsButton } from "./SwapEndsButton"
 
 const COMPONENT_EDGE_TYPE_OPTIONS: ReadonlyArray<EdgeTypeOption> = [
   { value: "ComponentDependency", label: "Dependency" },
@@ -16,19 +16,16 @@ const COMPONENT_EDGE_TYPE_OPTIONS: ReadonlyArray<EdgeTypeOption> = [
 export const ComponentEdgeEditPopover: React.FC<PopoverProps> = ({
   elementId,
 }) => {
-  const { getNode, updateEdgeData } = useReactFlow()
+  const { updateEdgeData } = useReactFlow()
 
   const edge = useReactiveEdge(elementId)
+  const sourceName = useReactiveNodeName(edge?.source, "Source")
+  const targetName = useReactiveNodeName(edge?.target, "Target")
   const { handleEdgeTypeChange, handleSwap } = useEdgePopOver(elementId)
 
   if (!edge) {
     return null
   }
-
-  const sourceNode = getNode(edge.source)
-  const targetNode = getNode(edge.target)
-  const sourceName = (sourceNode?.data?.name as string) ?? "Source"
-  const targetName = (targetNode?.data?.name as string) ?? "Target"
 
   const edgeData = edge.data as CustomEdgeProps | undefined
 
@@ -40,16 +37,7 @@ export const ComponentEdgeEditPopover: React.FC<PopoverProps> = ({
           updateEdgeData(elementId, { ...edge.data, [key]: value })
         }
         label="Control Flow"
-        sideElements={[
-          handleSwap && (
-            <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
-              <SwapHorizIcon
-                style={{ cursor: "pointer" }}
-                onClick={handleSwap}
-              />
-            </Box>
-          ),
-        ]}
+        sideElements={[handleSwap && <SwapEndsButton onClick={handleSwap} />]}
       />
 
       <EdgeTypeSelect
