+++ library/lib/components/popovers/edgePopovers/ReachabilityGraphEdgeEditPopover.tsx
@@ -2,9 +2,9 @@ import { Box } from "@mui/material"
 import { EdgeStyleEditor, TextField } from "@/components/ui"
 import { useReactFlow } from "@xyflow/react"
 import { CustomEdgeProps } from "@/edges/EdgeProps"
-import { SwapHorizIcon } from "@/components/Icon"
 import { useEdgePopOver, useReactiveEdge } from "@/hooks"
 import { PopoverProps } from "../types"
+import { SwapEndsButton } from "./SwapEndsButton"
 
 export const ReachabilityGraphEdgeEditPopover: React.FC<PopoverProps> = ({
   elementId,
@@ -28,16 +28,7 @@ export const ReachabilityGraphEdgeEditPopover: React.FC<PopoverProps> = ({
           updateEdgeData(elementId, { ...edge.data, [key]: value })
         }
         label="Reachability Arc"
-        sideElements={[
-          handleSwap && (
-            <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
-              <SwapHorizIcon
-                style={{ cursor: "pointer" }}
-                onClick={handleSwap}
-              />
-            </Box>
-          ),
-        ]}
+        sideElements={[handleSwap && <SwapEndsButton onClick={handleSwap} />]}
       />
 
       {/* Label update */}
