+++ library/lib/hooks/index.ts
@@ -5,6 +5,7 @@ export * from "./useNodeDragStop"
 export * from "./useNodeDrag"
 export * from "./useHandleOnResize"
 export * from "./useEdges"
+export * from "./useReactiveElement"
 export * from "./useEdgeToolbar"
 export * from "./useElementInteractions"
 export * from "./useGoToNextAssessment"
