+++ library/lib/hooks/useEdges.ts
@@ -1,28 +1,4 @@
-import { useReactFlow, useStore } from "@xyflow/react"
-
-/**
- * Reactive single-edge / single-node reads for edit popovers. Popovers must
- * subscribe to the element they edit (rather than reading getEdge/getNode
- * imperatively during render) so they re-render on data changes — swap, a
- * collaborator's edit, re-selection. Without this the React Compiler memoizes
- * the popover render and its controls show stale values.
- */
-export function useReactiveEdge(id: string) {
-  return useStore((s) => s.edgeLookup.get(id))
-}
-
-export function useReactiveNode(id: string) {
-  return useStore((s) => s.nodeLookup.get(id))
-}
-
-export function useReactiveNodeName(
-  id: string | undefined,
-  fallback: string
-): string {
-  return useStore(
-    (s) => (id && (s.nodeLookup.get(id)?.data?.name as string)) || fallback
-  )
-}
+import { useReactFlow } from "@xyflow/react"
 
 export function useEdgePopOver(id: string) {
   const reactFlow = useReactFlow()
